BstSorter()
Worked on the BstSorted to let it sort(). This method returns an ArrayList that can then be put into a BST, making a perfectly balanced BST (if there are (2^n)-1 elements in the list. I am still not sure what happens when you give it another amount of terms but I will be testing that.